Revert "Domantas"
