+++ .env
@@ -20,7 +20,7 @@ MAILER_URL=null://localhost
 # Format described at http://docs.doctrine-project.org/projects/doctrine-dbal/en/latest/reference/configuration.html#connecting-using-a-url
 # For an SQLite database, use: "sqlite:///%kernel.project_dir%/var/data.db"
 # Configure your db driver and server_version in config/packages/doctrine.yaml
-DATABASE_URL=mysql://root:@127.0.0.1:3306/testinis
+DATABASE_URL=mysql://root:root@127.0.0.1:3306/testinis
 ###< doctrine/doctrine-bundle ###
 
 ###> doctrine/doctrine-bundle ###
