+++ src/Controller/EmployeeController.php
@@ -34,8 +34,42 @@ class EmployeeController extends Controller
     }
 
 
+    /**
+     * @Route("/company/addJobs", name="addJobs")
+     */
+    public function addJob(Request $request, AuthorizationCheckerInterface $authChecker)
+    {
+        $em = $this->getDoctrine()->getManager();
+        $user = $this->getUser();
+        $newJob = new Job();
+        $form = $this->createForm(JobForm::class, $newJob);
+        $form->handleRequest($request);
 
+        if ($form->isSubmitted() && $form->isValid()) {
+            $newJob->setUser($user);
+            $em->persist($newJob);
+            $em->flush();
 
+            return $this->redirectToRoute('companyJobs');
+        }
+            return $this->render('Company/addJob.html.twig', array(
+                'registration_form' => $form->createView()));
+        }
 
+    /**
+     * @Route("/company/parts", name="partsPrice")
+     */
+    public function parts()
+    {
+        return $this->render('Company/parts.html.twig');
+    }
+
+    /**
+     * @Route("/company/jobs", name="companyJobs")
+     */
+    public function jobs()
+    {
+        return $this->render('Company/jobs.html.twig');
+    }
 
 }
