+++ src/Controller/HomeController.php
@@ -6,9 +6,6 @@ namespace App\Controller;
 use App\Events;
 use App\Entity\Orders;
 use App\Entity\Company;
-use App\Form\JobForm;
-use App\Entity\Job;
-use App\Entity\AdminJob;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
@@ -16,7 +13,7 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
-use Symfony\Component\HttpFoundation\Request;
+
 
 
 class HomeController extends AbstractController
@@ -24,7 +21,7 @@ class HomeController extends AbstractController
     /**
      * @Route("/homepage", name="homepage")
      */
-    public function index(AuthorizationCheckerInterface $authChecker, Request $request)
+    public function index(AuthorizationCheckerInterface $authChecker)
     {
         
         if ($authChecker->isGranted('ROLE_COMPANY'))
@@ -38,14 +35,6 @@ class HomeController extends AbstractController
             return $this->render('homepage.html.twig', [
 //                'pendingOrders' => $pendingOrders,
 //                'acceptedOrders' => $acceptedOrders
-            ]);
-        }
-        else if($authChecker->isGranted('ROLE_ADMIN')){
-            $em = $this->getDoctrine()->getManager();
-            $jobs = $em->getRepository(AdminJob::class)->findAll();
-
-            return $this->render('homepage.html.twig', [
-                'job' => $jobs,
             ]);
         }
         else
