+++ src/Entity/Administrator.php
@@ -19,6 +19,11 @@ class Administrator implements UserInterface
      */
     private $id;
 
+//    /**
+//     * @ORM\OneToOne(targetEntity="App\Entity\Administrator", mappedBy="userRole")
+//     */
+//    private $adminType;
+
     /**
      * @Assert\Length(
      *     min = 6,
@@ -71,7 +76,7 @@ class Administrator implements UserInterface
     private $service;
 
     /**
-     * @ORM\OneToMany(targetEntity="App\Entity\AdminJob", mappedBy="administrator")
+     * @ORM\OneToMany(targetEntity="App\Entity\Job", mappedBy="administrator")
      */
     private $job;
