+++ src/Repository/JobRepository.php
@@ -31,17 +31,4 @@ class JobRepository extends ServiceEntityRepository
         ;
     }
     */
-
-    public function findByCompany($id)
-    {
-        return $this->createQueryBuilder('c')
-            ->addSelect('r') // to make Doctrine actually use the join
-            ->leftJoin('c.jobName', 'r')
-            ->addSelect('u') // to make Doctrine actually use the join
-            ->leftJoin('r.company', 'u')
-            ->andwhere('c.company = :id')->setParameter('id', $id)
-            ->getQuery()
-            ->getResult();
-        ;
-    }
 }
