+++ templates/Company/jobs.html.twig
@@ -8,22 +8,22 @@
         <tr style="border-bottom: 3px solid #ddd;">
             <th width="120px">Pavadinimas</th>
             <th></th>
-            <th width="200px">Aprašymas</th>
-            <th></th>
             <th width="200px">Kaina</th>
             <th></th>
             <th width="200px">Trukmė</th>
+            <th></th>
+            <th width="200px">Aprašymas</th>
         </tr>
-        {% for job in jobs %}
-            <tr>
-                <td>{{ job.jobName }}</td>
-                <td></td>
-                <td>{{ job.description}}</td>
-                <td></td>
-                <td>{{ job.cost }}</td>
-                <td></td>
-                <td>{{ job.jobTime}}</td>
-            </tr>
-        {% endfor %}
+        {#{% for job in jobs %}#}
+            {#<tr>#}
+                {#<td>{{ job.jobName }}</td>#}
+                {#<td></td>#}
+                {#<td>{{ job.description}}</td>#}
+                {#<td></td>#}
+                {#<td>{{ job.cost }}</td>#}
+                {#<td></td>#}
+                {#<td>{{ job.jobTime}}</td>#}
+            {#</tr>#}
+        {#{% endfor %}#}
     </table>
 {% endblock %}
