--- templates/Registration/registrationJobs.html.twig
@@ -1,29 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body %}
-    <body>
-    <div class="container">
-        <div class="row">
-            <div class="col"></div>
-            <div class="col-5">
-                <h1>Darbu registracija</h1>
-                <div style="color: red">
-                    {#{% for message in error %}#}
-                    {#{{ message }}<br>#}
-                    {#{% endfor %}#}
-                </div>
-                <div id="default">
-                    {{ form_start(jobs_form) }}
-                    Darbo pavadinimas: {{ form_widget(jobs_form.jobsName, {'attr': {'class': 'AdminJob'}}) }}</br>
-                    <button class="btn btn-primary">Išsaugoti</button>
-                    {{ form_end(jobs_form) }}
-                </div>
-            </div>
-            <div class="col"></div>
-
-        </div>
-    </div>
-
-
-    </body>
-{%  endblock %}
